menu options

+++ src/com/skilldistillery/jets/CargoFreight.java
@@ -7,7 +7,7 @@ public class CargoFreight extends Jets implements FreightAircraft{
 	public CargoFreight() {
 	}
 	
-	public CargoFreight(String _model, double _speed, long _price, int _range, boolean _hasCargo ) {
+	public CargoFreight(String _model, Integer _speed, long _price, int _range, boolean _hasCargo ) {
 		super(_model, _speed, _price, _range);
 		hasCargo = _hasCargo;
 	}
